pipes v0.1
The list of upcoming pipes is working. Drops down on the left. Left
clicking will move the first upcoming pipe onto the board at the correct
position. Right click not implemented yet, but a handler has been put in
place. Lots of information stored in the Pipe object. Should be able to
be used to deal with connections/water filled.